create "deprecated" folder for unused scripts
Some of these scripts may be useful later but will likely need to be refactored (post_annotation scripts in particular still need to be worked back into pipeline)
Some others are now redundant and do not work smoothly in existing notebooks.
This folder should eventually be removed from the repo.