change permissions when creating new folders
Because the notebooks creates folders as root, users can't modify them from a file explorer,
(which is the easiest/most straightforward way for me to use the processed images). After folders are created,
the permissions are changed to "create and delete files" allowed for owner, groups, and others. This is done
with the understanding that folders that may be mistakenly deleted or edited can be remade or redownloaded;
it is more useful to allow the folders to be edited than it is useful to protect them. Although intermediate
files may be useful to reference for troubleshooting during the process, most of the images generated during
this process are not useful to keep once training data has been made.