2 d options (#22)
* initial commit of new notebooks for 2D data pipeline

2D data (no time or z dimension data needs to be preserved) doesn't need to be annotated as montages,
so it doesn't make sense to use the same notebook for preparing this data for crowdsourcing annotations,
or using the annotations (not all functions needed, different file structure)

* add 2D variable to overlapping chopper

-removes montage steps from pre- and post-processing notebooks
-changes overlapping chopper so that if 2D mode is toggled on, all chopped images are saved in same folder
-slightly modifies instructions of notebooks

* change 2D chopper subimage names

subimages saved by chopper don't need to have frame number in file name if they are 2D
minor changes to notebook descriptions

* minor changes to descriptions

* add "make_npz" step to end of pipeline

adds make_training_data to end of notebook,
uses variables from earlier in notebook and in json logs for make_training_data,
descriptions added and updated throughout notebook

* added argument to montage chopper to specify where to find annotated montages

user-defined annotations folder instead of default folder named "annotations"

* change permissions when creating new folders

Because the notebooks creates folders as root, users can't modify them from a file explorer,
(which is the easiest/most straightforward way for me to use the processed images). After folders are created,
the permissions are changed to "create and delete files" allowed for owner, groups, and others. This is done
with the understanding that folders that may be mistakenly deleted or edited can be remade or redownloaded;
it is more useful to allow the folders to be edited than it is useful to protect them. Although intermediate
files may be useful to reference for troubleshooting during the process, most of the images generated during
this process are not useful to keep once training data has been made.

* fix indentation

* add support for dealing with broken_links from figure eight job report

keeps track of broken_links so user can inspect and resubmit rows that failed
also keeps track of whether the broken_link annotations have successful annotations later in the job
so that if reuploaded broken_link images are finished successfully, no csv for reuploading is saved

* update description for download annotations step

* intial commit of overlapping_stitcher

includes overlapping_stitcher core function in post_annotation scripts, updated __init__,
and a wrapper for 2D image stitching in the post-annotation notebook

* fix typo

* add is_2D option to contrast_adjustment for different filename formats

* add features to contrast adjustment

-adds 2D vs 3D naming option
-changes arguments of adjust_folder to be more concise
-adds json logging of settings
-adds v_min and v_max settings to single-image contrast adjustment

* montage maker no longer uses nested subfolders

-fewer differences between 2D and 3D pipelines helps make functions more general
-small naming changes to overlapping chopper
-montage maker input now based on info in filenames, not on position of image in a list
-montage maker suppresses low contrast image warning

* add support for including prev and next images in fig8 job

-images named with 3D file naming can be displayed to annotators with
previous and next images for context
-previous and next image urls generated in aws upload script,
optionally included in csv
-notebooks updated to function with new function args

* update how images are stitched together

-uses numpy functions to speed this up
-uses cells that overlap the most, not the first pair of cells that overlap each other
-no longer overwrites annotations with background when only one image has an annotation in overlap region
-relabels images at end

* add frame_offset to overlapping_chopper

if separating a movie into different parts for annotation, helps name images with correct frame/image number
record frame_offset and number of images chopped in json log

* add wrapper for stitcher, small changes to core stitcher

-switch to using saved indices when applicable
-use zeros if a sub image can't be loaded

* minor updates to pre-annotation 2D

-use imageio imread instead of skimage imread
-add frame_offset input to overlapping_chopper
-fix output of aws_upload
-add more fig8 job numbers to copy

* add make_training_data functionality to post-processing notebook

-add option to chop raw images to match annotation size
-update option to stitch annotations together (now uses overlapping_stitcher_folder)
-move appropriate folders of images into training folder
-optional additional subfolder creation to work with make_training_data
-add code block to setting variables for and running make_training_data

* add deepcell-tf data utils

deepcell-tf data utils are more up to date than many deepcell-toolbox data utils, so added for reference/possible use
updated init correspondingly
changed name of 3D post-processing notebook

* new make_training_data functions, more appropriate for post-annotation pipeline

-update init to include new file
-2D_make_training_data uses all images in provided raw and annotation folders
-make_training_data uses folders to determine images to load, does not use regex, images in folders must be in correct order when sorted
-change optional code blocks in notebook for rearranging images in training folder
-change make_training_data code blocks in notebook to use new functions appropriately

* fix typo

* change npz formation and file rearrangment into subfolders

-montage-related scripts use image names more similar to those used in non-montaged pipeline
-putting chopped images into nested subfolders now happens in separate set of code blocks in notebook, not within montage chopper function
-npz formation uses all montages (instead of just one part at a time), where each montage is its own batch
-uses post_annotation_make_training_data

* update documentation of pre-annotation functions

* add .jpg to file extensions in get_img_names

* update documentation for utils

* add reshape_matrix and update documentation

* update documentation