set up caliban pre-annotation pipeline
Create pre-annotation pipeline to set up Figure Eight jobs with Caliban tool.

Caliban can access data files by taking the filepath appended to its url and requesting to view the file from its S3 bucket. The pipeline has been updated to upload .npz/.trk files to a predefined input bucket and subfolder path. The path is then taken to construct the unique caliban.deepcell.org url to access the file on the browser tool. The url for each file is added to a .CSV file that is stored in the same directory as the input data folder. The pipeline finishes by setting up a Figure Eight job to launch for data curation.