create npz preprocessing functions (reshape, slice, relabel)
These functions are meant to be used to prepare npz files for crowd-sourced
annotation with Caliban. These focus on turning large npzs into manageable
pieces by reshaping and slicing them. Functions to relabel npzs are included
so that annotators have consistent labeling in the uncorrected npzs. Also
included are functions to wrap slicing and relabeling functions, so that they
can easily be run on folders of npz files.

npz_preprocessing.py added to __init__.py